refactor: move page components to pages/ folder
